Fleshed out more UI stuff and reworked log code
Added more info to the UI including actual output to the (now scrollable) info text box. I switched to using ttk instead of tk widgets so they have a more modern feel

I also reworked the original route copier code to work more as a file watcher and event callback executor. Basically anytime a log entry is made (or when the program first reads the logs), the code gets what event it was and it gathers a list of callbacks (registered externally) that are waiting for those events. It then executes any that are listening for that particular event and also executes any listeners who want all events sent to them.